waypoint initial commit

+++ database/routes/src/main/java/ua/vertex/waypoint/DAO/WayPointDAO.java
@@ -8,9 +8,9 @@ import java.util.List;
  * Created by Vasyl on 18/05/2016.
  */
 public interface WayPointDAO {
-    int createWayPoint(WayPoint wayPoint);
-    WayPoint readWayPoint(int id);
-    void updateWayPoint(int id, WayPoint wayPoint);
-    void deleteWayPoint(int id);
+    long createWayPoint(WayPoint wayPoint);
+    WayPoint readWayPoint(long id);
+    void updateWayPoint(long id, WayPoint wayPoint);
+    void deleteWayPoint(long id);
     List<WayPoint> getSortedWayPointsForRoute(long routeId);
 }
